refactor: move values into different files
